Gestion de la mise à jour du solde client après transaction via Kafka

+++ Client_Service/Client_Service/src/main/java/com/example/Client_Service/repo/ClientRepo.java
@@ -3,5 +3,5 @@ package com.example.Client_Service.repo;
 import com.example.Client_Service.Model.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientRepo extends JpaRepository<Client,Integer> {
+public interface ClientRepo extends JpaRepository<Client,Long> {
 }
